add virtual env to ignored files
observation: I read in a site that it's a bad practice to put these
files into a .gitignore. However, the site also said that it's not good
to add these files to the repository. So I'm still a bit confused about
what to do with them. And since adding them to the repository seems to
be the worst option, I'm adding them to the ignored files.