transform notebook into proper pipeline files
create a class to load the dataset from the folder and return a pandas
DataFrame.

create a class to train the model.

create a class to test the model (it actually use a pre-trained model to
make predictions, so maybe its name will be changed in the near future).

delete the old Initial_File.ipynb notebook because now we can just use
the classes created above.

create a Test.ipynb notebook to test the classes.